+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add provider independent controllers to manage labeling and setting owner references in other provider dependent objects.
 - Export container logs for e2e tests to azure analytics.
 - Enable persistent volume `expansion` support in the default `Storage Classes`.
+- Added to all VMSSes the tags needed by cluster autoscaler.
 
 ### Changed